LC Testimonails Slider
LC Testimonial Slider WordPress plugin is a great tool to show your
clients testimonials. You can display testimonials in any post or page
content
simply by pasting shortcode in it in a clean, responsive and beautiful
Slider format. The "LC Testimonials" are a Custom Post Type. You can
enter the
Name, Company, Website in the Custom Fields for the Testimonial and the
actual Testimonial text in the "Description" and image as featured
image.